Split code into different files
