+++ src/main/scala/bot/imgur/Service.scala
@@ -1,0 +1,7 @@
+package bot.imgur
+
+import scala.concurrent.Future
+
+trait Service {
+  def getRandomCat: Future[String]
+}
