+++ src/main/scala/bot/imgur/api/Response.scala
@@ -1,0 +1,7 @@
+package bot.imgur.api
+
+case class Response(data: List[Data])
+
+case class Data(images: List[InnerData])
+
+case class InnerData(link: String)
