+++ src/main/scala/bot/imgur/random/RandomRest.scala
@@ -1,0 +1,7 @@
+package bot.imgur.random
+
+import scala.util.Random
+
+object RandomRest extends Randomize {
+  override def randomElem[T](ls: List[T]): T = Random.shuffle(ls).head
+}
