+++ src/main/scala/bot/imgur/random/Randomize.scala
@@ -1,0 +1,5 @@
+package bot.imgur.random
+
+trait Randomize {
+  def randomElem[T](ls: List[T]): T
+}
